Auth: Delivering the newest auth flow for VSTS using AAD to verify the user's identity; this method leverages the static tab to keep everything in Teams; due to change the ComposeExtensionSettingsPopUp was added for later
